backend/requirements.txt
Django==5.0.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
celery==5.4.0
redis==5.0.8
python-dotenv==1.0.1
requests==2.32.3
shapely==2.0.6
sentry-sdk==2.13.0
geopy==2.4.1
gunicorn==22.0.0
whitenoise==6.7.0
